Refine cleanup dispatcher delivery handling
